+++ pydbtools/utils.py
@@ -1,10 +1,15 @@
-from typing import Tuple
+from typing import Tuple, Optional
 import os
 import re
 import sqlparse
+import sql_metadata
 import inspect
 import boto3
-from botocore.credentials import InstanceMetadataProvider, InstanceMetadataFetcher
+from botocore.credentials import (
+    InstanceMetadataProvider,
+    InstanceMetadataFetcher,
+)
+
 
 # Set pydbtool params - if you were so inclined to change them
 bucket = "mojap-athena-query-dump"
@@ -12,6 +17,18 @@ temp_database_name_prefix = "mojap_de_temp_"
 aws_default_region = os.getenv(
     "AWS_DEFAULT_REGION", os.getenv("AWS_REGION", "eu-west-1")
 )
+
+
+def _set_aws_session_name():
+    if not os.getenv("AWS_ROLE_SESSION_NAME"):
+        os.environ["AWS_ROLE_SESSION_NAME"] = _get_role_name_from_env()
+
+
+def _get_role_name_from_env() -> str:
+    aws_role_arn = os.getenv("AWS_ROLE_ARN")
+    if aws_role_arn is None:
+        raise EnvironmentError("AWS_ROLE_ARN was not found in env")
+    return aws_role_arn.split("/")[-1]
 
 
 def _set_region_name(region_name: str):
@@ -48,12 +65,11 @@ def check_temp_query(sql: str):
         )
 
 
-def clean_query(sql: str, fmt_opts: dict = None) -> str:
+def clean_query(sql: str, fmt_opts: Optional[dict] = None) -> str:
     """
     removes trailing whitespace, newlines and final
     semicolon from sql for use with
     sqlparse package
-
     Args:
         sql (str): The raw SQL query
         fmt_opts (dict): Dictionary of params to pass to sqlparse.format.
@@ -61,9 +77,11 @@ def clean_query(sql: str, fmt_opts: dict = None) -> str:
     Returns:
         str: The cleaned SQL query
     """
+    if fmt_opts is None:
+        fmt_opts = {}
+    fmt_opts["strip_comments"] = True
+    sql = sqlparse.format(sql, **fmt_opts)
     sql = " ".join(sql.splitlines()).strip().rstrip(";")
-    if fmt_opts:
-        sql = sqlparse.format(sql, **fmt_opts)
     return sql
 
 
@@ -79,23 +97,26 @@ def replace_temp_database_name_reference(sql: str, database_name: str) -> str:
     Returns:
         str: The new SQL query which is sent to Athena
     """
-    # check query is valid and clean
 
-    parsed = sqlparse.parse(clean_query(sql))
+    parsed = sqlparse.parse(sql)
     new_query = []
     for query in parsed:
         check_temp_query(str(query))
-        for word in str(query).strip().split(" "):
-            if "__temp__." in word.lower():
-                word = word.lower().replace("__temp__.", f"{database_name}.")
-            new_query.append(word)
-        if ";" not in new_query[-1]:
-            last_entry = new_query[-1] + ";"
-        else:
-            last_entry = new_query[-1]
-        del new_query[-1]
-        new_query.append(last_entry)
-    return " ".join(new_query)
+        # Get all the separated tokens from subtrees
+        fq = list(query.flatten())
+        # Join them back together replacing __temp__
+        # where necessary
+        new_query.append(
+            "".join(
+                re.sub(
+                    "^__temp__", database_name, str(word), flags=re.IGNORECASE
+                )
+                for word in fq
+            )
+        )
+    # Strip output for consistency, different versions of sqlparse
+    # treat a trailing newline differently
+    return "".join(new_query).strip()
 
 
 def get_user_id_and_table_dir(
@@ -105,7 +126,9 @@ def get_user_id_and_table_dir(
     region_name = _set_region_name(region_name)
 
     if boto3_session is None:
-        boto3_session = get_boto_session(force_ec2=force_ec2, region_name=region_name)
+        boto3_session = get_boto_session(
+            force_ec2=force_ec2, region_name=region_name
+        )
 
     sts_client = boto3_session.client("sts")
     sts_resp = sts_client.get_caller_identity()
@@ -122,15 +145,45 @@ def get_database_name_from_userid(user_id: str) -> str:
     return unique_db_name
 
 
+def get_database_name_from_sql(sql: str) -> str:
+    """
+    Obtains database name from SQL query for use
+    by awswrangler.
+
+    Args:
+        sql (str): The raw SQL query as a string
+
+    Returns:
+        str: The database table name
+    """
+
+    for table in sql_metadata.Parser(sql).tables:
+        # Return the first database seen in the
+        # form "database.table"
+        xs = table.split(".")
+        if len(xs) > 1:
+            return xs[0]
+
+    # Return default in case of failure to parse
+    return None
+
+
 def get_boto_session(
-    force_ec2: bool = False, region_name: str = None,
+    force_ec2: bool = False,
+    region_name: str = None,
 ):
+    # Check for new platform authentication
+    if os.getenv("AWS_ROLE_ARN") is not None:
+        _set_aws_session_name()
+
     region_name = _set_region_name(region_name)
 
     kwargs = {"region_name": region_name}
     if force_ec2:
         provider = InstanceMetadataProvider(
-            iam_role_fetcher=InstanceMetadataFetcher(timeout=1000, num_attempts=2)
+            iam_role_fetcher=InstanceMetadataFetcher(
+                timeout=1000, num_attempts=2
+            )
         )
         creds = provider.load().get_frozen_credentials()
         kwargs["aws_access_key_id"] = creds.access_key
@@ -150,6 +203,8 @@ def get_boto_client(
     region_name = _set_region_name(region_name)
 
     if boto3_session is None:
-        boto3_session = get_boto_session(force_ec2=force_ec2, region_name=region_name)
+        boto3_session = get_boto_session(
+            force_ec2=force_ec2, region_name=region_name
+        )
 
     return boto3_session.client(client_name)